Moved to component folder. GA event
