+++ src/components/StablecoinsDataDownloadBtn/StablecoinsDataDownloadBtn.module.scss
@@ -1,5 +1,6 @@
 .button {
   position: relative;
+  margin-left: 15px !important;
 }
 
 .link {
